+++ jjava-launcher/src/main/java/org/dflib/jjava/launcher/KernelLauncher.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.logging.Logger;
 
 /**
  * The launcher responsible for creating a new process to run the JJava kernel.
@@ -13,6 +14,15 @@ import java.util.List;
  */
 public class KernelLauncher {
 
+    // Since "launcher" is kind of a part of the "distro", we are using JUL directly,
+    // unlike the rest of the framework that uses SLF4J, and is only linked to JUL in the "distro"
+    private static final Logger LOGGER = Logger.getLogger("KernelLauncher");
+
+    // More JUL stuff... This is copied verbatim from JJava class to mimic Jupyter log format
+    // TODO: does it make sense to pass this property to the child JJava process via -D,
+    //   to avoid duplicating this logic?
+    private static final String JUL_JUPYTER_LOG_FORMAT = "[%4$.1s %1$tF %1$tT.%1$tL %3$s] %5$s%n";
+
     private static final String JJAVA_JVM_OPTS = "JJAVA_JVM_OPTS";
 
     private final List<String> args;
@@ -21,6 +31,12 @@ public class KernelLauncher {
         this.args = args;
     }
 
+    public static void main(String[] args) {
+        System.setProperty("java.util.logging.SimpleFormatter.format", JUL_JUPYTER_LOG_FORMAT);
+        KernelLauncher launcher = new KernelLauncher(Arrays.asList(args));
+        int exitCode = launcher.launchKernel();
+        System.exit(exitCode);
+    }
 
     /**
      * Launches a new process to run the JJava kernel with the given arguments.
@@ -33,7 +49,7 @@ public class KernelLauncher {
      */
     public int launchKernel() {
         List<String> command = buildCommand(args);
-        System.out.println("Running JJava Kernel with args: " + command);
+        LOGGER.info(() -> "Starting kernel: " + command);
         try {
             ProcessBuilder pb = new ProcessBuilder(command);
             pb.inheritIO();
@@ -47,11 +63,11 @@ public class KernelLauncher {
 
     private void addShutdownHook(Process kernel) {
         Runtime.getRuntime().addShutdownHook(new Thread(() -> {
-            System.out.println("Launcher is shutting down, terminating the kernel...");
+            LOGGER.fine("Terminating kernel...");
             kernel.destroy();
             try {
                 kernel.waitFor();
-                System.out.println("Kernel terminated.");
+                LOGGER.info("Kernel terminated.");
             } catch (InterruptedException e) {
                 throw new RuntimeException(e);
             }
@@ -85,12 +101,14 @@ public class KernelLauncher {
     }
 
     private String buildErrorMessage(List<String> args) {
-        if (args.isEmpty()) {
-            return String.format("Missing arguments: <kernel_path> <connection_file>%n");
+
+        switch (args.size()) {
+            case 0:
+                return "Missing arguments: <kernel_path> <connection_file>%n";
+            case 1:
+                return String.format("Missing arguments: %s <connection_file>%n", args.get(0));
+            default:
+                return String.format("Arguments provided: %s%n", String.join(", ", args));
         }
-        if (args.size() < 2) {
-            return String.format("Missing arguments: %s <connection_file>%n", args.get(0));
-        }
-        return String.format("Arguments provided: %s%n", String.join(", ",args));
     }
 }