+++ RELEASE-NOTES.md
@@ -19,6 +19,7 @@
 * #90 Deprecate Ivy artifact syntax for %maven magic
 * #93 Extensions / kernel lifecycle
 * #99 Get rid of SLF4J transitive dependency
+* #100 Organize internal logging around SLF4J and JUL
 
 ## 1.0-a5